Second task adding a condition

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,4 +1,6 @@
 #killing a process
 file {'cwd':
-  command => 'pkill -f killmenow'
+  command     => 'pkill -f killmenow'
+  refreshonlu => true
+  onlyif      => 'pgrep -f killmenow'
 }
